+++ collect_app/build.gradle
@@ -274,7 +274,7 @@ dependencies {
     implementation project(':db')
     implementation project(':open-rosa')
     implementation project(':mobile-device-management')
-    implementation project(':experimental')
+    implementation project(':timedgrid')
 
     if (getSecrets().getProperty('MAPBOX_DOWNLOADS_TOKEN', '') != '') {
         implementation project(':mapbox')
