+++ collect_app/src/main/java/org/odk/collect/android/utilities/Appearances.kt
@@ -91,8 +91,6 @@ object Appearances {
     const val MASKED = "masked"
     const val COUNTER = "counter"
     const val MULTILINE = "multiline"
-
-    // Experimental
     const val TIMED_GRID = "timed-grid"
 
     // Get appearance hint and clean it up so it is lower case, without the search function and never null.
