+++ timedgrid/src/main/AndroidManifest.xml
