+++ timedgrid/src/main/java/org/odk/collect/timedgrid/NavigationAwareWidget.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.experimental.timedgrid
+package org.odk.collect.timedgrid
 
 import androidx.fragment.app.DialogFragment
