+++ timedgrid/src/main/java/org/odk/collect/timedgrid/PausableCountDownTimer.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.experimental.timedgrid
+package org.odk.collect.timedgrid
 
 import android.os.CountDownTimer
