+++ timedgrid/src/main/java/org/odk/collect/timedgrid/TimedGridRenderer.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.experimental.timedgrid
+package org.odk.collect.timedgrid
 
 import android.view.LayoutInflater
 import android.view.View
