+++ timedgrid/src/main/java/org/odk/collect/timedgrid/TimedGridState.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.experimental.timedgrid
+package org.odk.collect.timedgrid
 
 // Lifecycle states for the timed grid widget.
 enum class TimedGridState {
