+++ timedgrid/src/main/res/layout/timed_grid_item_row.xml
