+++ timedgrid/src/main/res/values/colors.xml
