+++ timedgrid/src/main/res/values/dimens.xml
