+++ timedgrid/src/main/res/values/strings.xml
