+++ timedgrid/src/main/res/values/styles.xml
